Update 00040000000EC700.txt
Etrian Odyssey Untold. Noticed some codes didn't work in Classic mode because the addresses only pointed to the 5 Story mode characters. Did a hex calculation and found the addresses that made up the first 15 guild slots for Classic Mode. The offset is 180 for each unit after in hex, so try that with some of the other Boost/Stats 99/Level 70 after battle codes.